Actualizando el main para el backend

+++ Taller_TIC1_Backend/Models/DTOs/FacturaCreateDTO.cs
@@ -4,9 +4,9 @@ namespace Taller_TIC1_Backend.Models.DTOs
     {
         public int IdOrdenTrabajo { get; set; }
         public DateTime FechaEmision { get; set; } = DateTime.Now;
-        public float Subtotal { get; set; }
-        public float Impuestos { get; set; }
-        public float Total { get; set; }
+        public decimal Subtotal { get; set; }
+        public decimal Impuestos { get; set; }
+        public decimal Total { get; set; }
         public string Estado { get; set; } = "Pendiente";
     }
 }
