+++ Taller_TIC1_Backend/Models/DTOs/RepuestoCantidadDTO.cs
@@ -5,7 +5,7 @@
         public int IdRepuesto { get; set; }
         public string NombreRepuesto { get; set; } = string.Empty;
         public int Cantidad { get; set; }
-        public float PrecioUnitario { get; set; }
-        public float Subtotal { get; set; }
+        public decimal PrecioUnitario { get; set; }
+        public decimal Subtotal { get; set; }
     }
 }
